Cosmetic changes only, mostly comments.
Changed the argument list for one of the log files
to take c language primitives instead of the derived
types Geometrydata and Analysisdata.  The short story
here is that using derived types makes more work
in the long term, and is harder to test.  The long
answer is a couple of pages of explanation.